requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==3.0.4
chardet==3.0.4
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
deep-translator==1.11.4
Django==4.1
django-cors-headers==4.4.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pgeocode==0.5.0
pillow==10.4.0
prometheus_client==0.20.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.20.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==6.4.post2
